requirements.txt
opencv-python-headless==4.7.0.68
Pillow==9.5.0
pytesseract==6.3.0
python-dotenv==1.0.0
transformers==4.31.0
UnstructuredTextPartioner==0.2.1
langchain==0.0.40
torch==2.0.1
torchaudio===2.0.2
torchvision===0.15.2
faiss-cpu==1.8.0
tiktoken==0.5.0
openai-whisper==20230315
openai==0.27.8
streamlit==1.24.0